Separate UserRowsLabel from UserRows
