+++ TweetNest/Shared/UserRows.swift
@@ -11,7 +11,7 @@ import OrderedCollections
 import TweetNestKit
 
 extension String {
-    fileprivate var displayUserID: String {
+    var displayUserID: String {
         Int64(self).flatMap { "#\($0.twnk_formatted())" } ?? "#\(self)"
     }
 }
@@ -24,8 +24,6 @@ struct UserRows<Icon: View>: View {
     let searchQuery: String
 
     let icon: Icon?
-
-    @StateObject private var latestUserDetailsFetchedResultsController: FetchedResultsController<UserDetail>
 
     private var filteredUserIDs: OrderedSet<String> {
         guard searchQuery.isEmpty == false else {
@@ -46,55 +44,24 @@ struct UserRows<Icon: View>: View {
         ])
         fetchRequest.propertiesToFetch = ["id"]
 
-        filteredUserIDs.append(contentsOf: (try? viewContext.fetch(fetchRequest).compactMap { $0["id"] as? String }) ?? [])
+        let fetchResults = try? viewContext.fetch(fetchRequest)
+        filteredUserIDs.append(contentsOf: fetchResults?.compactMap { $0["id"] as? String } ?? [])
 
         return userIDs.intersection(filteredUserIDs)
     }
 
     var body: some View {
-        let latestUserDetailsByUserID = Dictionary(grouping: latestUserDetailsFetchedResultsController.fetchedObjects) {
-            $0.user?.id
-        }
-
         ForEach(filteredUserIDs, id: \.self) { userID in
-            let latestUserDetail = latestUserDetailsByUserID[userID]?.last
-
-            Label {
+            NavigationLink {
+                UserView(userID: userID)
+                    .environment(\.account, account)
+            } label: {
                 Label {
-                    NavigationLink {
-                        UserView(userID: userID)
-                            .environment(\.account, account)
-                    } label: {
-                        if let latestUserDetail = latestUserDetail {
-                            TweetNestStack {
-                                Text(verbatim: latestUserDetail.name ?? userID.displayUserID)
-                                    .lineLimit(1)
-
-                                if let username = latestUserDetail.username {
-                                    Text(verbatim: "@\(username)")
-                                        .lineLimit(1)
-                                        .layoutPriority(1)
-                                        .foregroundColor(Color.gray)
-                                }
-                            }
-                        } else {
-                            Text(verbatim: userID.displayUserID)
-                                .lineLimit(1)
-                        }
-                    }
+                    UserRowsLabel(userID: userID)
                 } icon: {
-                    if let latestUserDetail = latestUserDetail {
-                        ProfileImage(profileImageURL: latestUserDetail.profileImageURL)
-                            .frame(width: 24, height: 24)
-                    }
+                    icon
                 }
-                #if os(watchOS)
-                .labelStyle(.titleOnly)
-                #endif
-            } icon: {
-                icon
             }
-            .accessibilityLabel(Text(verbatim: latestUserDetail?.name ?? userID.displayUserID))
         }
     }
 
@@ -102,23 +69,6 @@ struct UserRows<Icon: View>: View {
         self.userIDs = OrderedSet(userIDs)
         self.searchQuery = searchQuery
         self.icon = icon
-
-        self._latestUserDetailsFetchedResultsController = StateObject(
-            wrappedValue: FetchedResultsController<UserDetail>(
-                fetchRequest: {
-                    let fetchRequest = UserDetail.fetchRequest()
-                    fetchRequest.predicate = NSPredicate(format: "user.id IN %@", Array(userIDs))
-                    fetchRequest.sortDescriptors = [
-                        NSSortDescriptor(keyPath: \UserDetail.user?.modificationDate, ascending: false),
-                        NSSortDescriptor(keyPath: \UserDetail.creationDate, ascending: false),
-                    ]
-                    fetchRequest.relationshipKeyPathsForPrefetching = ["user"]
-
-                    return fetchRequest
-                }(),
-                managedObjectContext: TweetNestApp.session.persistentContainer.viewContext
-            )
-        )
     }
 
     init<S>(userIDs: S, searchQuery: String = "", @ViewBuilder icon: () -> Icon) where S: Sequence, S.Element == String {
